requirements.txt
pillow
requests
scratchattach
gunicorn
Flask